[Gold III] Title: 괄호 추가하기, Time: 124 ms, Memory: 14376 KB -BaekjoonHub

+++ 백준/Gold/16637. 괄호 추가하기/괄호 추가하기.java
@@ -2,7 +2,7 @@ import java.io.*;
 
 public class Main {
 	static int N;
-	static long res;
+	static int res;
 	static String[] str;
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
@@ -13,21 +13,21 @@ public class Main {
 		System.out.print(res);
 	}
 	
-	public static void back(int depth, long sum) {
+	public static void back(int depth, int sum) {
 		if (depth >= N/2) {
 			res = Math.max(res, sum);
 			return;
 		}
 		int now = 2 * depth;
-		long a = Long.parseLong(str[now + 2]);
+		int a = Integer.parseInt(str[now + 2]);
 		back(depth+1, cal(sum, a, str[now+1]));
 		if (now + 4 < N) {
-			long b = Long.parseLong(str[now + 4]);
+			int b = Integer.parseInt(str[now + 4]);
 			back(depth+2, cal(sum, cal(a, b, str[now+3]), str[now+1]));
 		}
 	}
 	
-	public static long cal(long a, long b, String c) {
+	public static int cal(int a, int b, String c) {
 		if (c.equals("+")) return a + b;
 		else if (c.equals("-")) return a - b;
 		return a * b;
